# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Service API",
        "description": "API documentation for the translation service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1": {},
        "/api/v1/login": {
            "post": {
                "summary": "User login",
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get all translations",
                "operationId": "fe964886b9e575d32578d804e0aa968b",
                "responses": {
                    "200": {
                        "description": "List of translations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "operationId": "2415590069171f9072e07aa15aad9ffc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "welcome_message"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Welcome to our website"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created successfully"
                    }
                }
            }
        },
        "/api/v1/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export all translations as JSON",
                "operationId": "1c4c422bcac4cac5877f8b0b86f37f8f",
                "responses": {
                    "200": {
                        "description": "Successfully exported translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "en": {
                                        "welcome_message": "Welcome to our website",
                                        "goodbye_message": "Goodbye and see you soon"
                                    },
                                    "fr": {
                                        "welcome_message": "Bienvenue sur notre site",
                                        "goodbye_message": "Au revoir et à bientôt"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Translations",
            "description": "Translations"
        }
    ]
}